+++ .gitignore
@@ -107,7 +107,7 @@ venv.bak/
 .mypy_cache/
 
 # db
-*.db
+# *.db
 
 # IDE
 .vscode
