--- flask_mysql/Procfile
@@ -1,1 +1,0 @@
-web: gunicorn main:app
