--- flask_mysql/app/errors.py
@@ -1,13 +1,0 @@
-from flask import jsonify, make_response
-from app import app
-
-
-# HTTP error handling
-@app.errorhandler(404)
-def not_found(error):
-    return make_response(jsonify({'error': 'Not found'}), 404)
-
-
-@app.errorhandler(400)
-def bad_request(error):
-    return make_response(jsonify({'error': 'Bad request'}), 400)
