--- flask_mysql/nltk.txt
@@ -1,1 +1,0 @@
-punkt
