+++ main.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python
+
+import os
+from app import app
+
+# Run application
+if __name__ == '__main__':
+    app.run()
