+++ requirements.txt
@@ -1,0 +1,20 @@
+Click==7.0
+decorator==4.3.2
+Flask==1.0.2
+Flask-SQLAlchemy==2.3.2
+flask-cors==3.0.7
+gunicorn==19.9.0
+itsdangerous==1.1.0
+Jinja2==2.10
+MarkupSafe==1.1.1
+networkx==2.2
+nltk==3.4
+numpy==1.16.2
+pandas==0.24.1
+PyMySQL==0.9.3
+python-dateutil==2.8.0
+pytz==2018.9
+singledispatch==3.4.0.3
+six==1.12.0
+SQLAlchemy==1.3.1
+Werkzeug==0.14.1
